Created GameStreamService and interface to house gameStreamMethods
