+++ the-backfield/Interfaces/IPlayService.cs
@@ -9,7 +9,6 @@ namespace TheBackfield.Interfaces
         Task<ResponseDTO<Play>> GetSinglePlayAsync(int playId, string sessionKey);
         Task<ResponseDTO<Play>> CreatePlayAsync(PlaySubmitDTO playSubmit);
         Task<ResponseDTO<Play>> UpdatePlayAsync(PlaySubmitDTO playSubmit);
-        Task<List<PlaySegmentDTO>> GetPlaySegmentsAsync(int playId);
         Task<ResponseDTO<Play>> DeletePlayAsync(int playId, string sessionKey);
     }
 }
