+++ the-backfield/Services/GameService.cs
@@ -3,32 +3,24 @@ using TheBackfield.DTOs;
 using TheBackfield.Interfaces;
 using TheBackfield.Models;
 using TheBackfield.Utilities;
-using System.Xml.Linq;
-using Microsoft.AspNetCore.Routing.Patterns;
 
 namespace TheBackfield.Services;
 
 public class GameService : IGameService
 {
     private readonly IGameRepository _gameRepository;
-    private readonly IPlayRepository _playRepository;
     private readonly ITeamRepository _teamRepository;
     private readonly IUserRepository _userRepository;
-    private readonly IPlayService _playService;
 
     public GameService(
         IGameRepository gameRepository,
-        IPlayRepository playRepository,
         ITeamRepository teamRepository,
-        IUserRepository userRepository,
-        IPlayService playService
+        IUserRepository userRepository
         )
     {
         _gameRepository = gameRepository;
-        _playRepository = playRepository;
         _teamRepository = teamRepository;
         _userRepository = userRepository;
-        _playService = playService;
     }
 
     public async Task<ResponseDTO<Game>> CreateGameAsync(GameSubmitDTO gameSubmit)
@@ -89,156 +81,6 @@ public class GameService : IGameService
         User? user = await _userRepository.GetUserBySessionKeyAsync(sessionKey);
         Game? game = await _gameRepository.GetSingleGameAsync(gameId);
         return SessionKeyClient.VerifyAccess(sessionKey, user, game);
-    }
-
-    public async Task<GameStreamDTO?> GetGameStreamAsync(int gameId)
-    {
-        Game? game = await _gameRepository.GetSingleGameAllStatsAsync(gameId);
-        if (game == null)
-        {
-            return null;
-        }
-
-        int down = 0;
-        int? toGain = null;
-        int? fieldPositionStart = null;
-        int? nextTeamId = null;
-        PlayAsSegmentsDTO? lastPlay = null;
-
-        int currentPlayId = game.Plays.SingleOrDefault(p => !game.Plays.Any(gp => gp.PrevPlayId == p.Id))?.Id ?? 0;
-        Play? currentPlay = await _playRepository.GetSinglePlayAsync(currentPlayId);
-        var (homeTeamScore, awayTeamScore) = StatClient.ParseScore(game);
-
-        if (currentPlay != null)
-        {
-            (down, toGain, fieldPositionStart, nextTeamId) = StatClient.ParseNextFieldPosition(currentPlay, game.HomeTeamId, game.AwayTeamId);
-            lastPlay = new PlayAsSegmentsDTO(currentPlay, await _playService.GetPlaySegmentsAsync(currentPlayId));
-        }
-
-        int? clockStart = null;
-        int? gamePeriod = null;
-        int playCheckId = currentPlay?.Id ?? 0;
-        do
-        {
-            Play? checkPlay = game.Plays.SingleOrDefault(p => p.Id == playCheckId);
-            if (checkPlay == null)
-            {
-                clockStart = game.PeriodLength;
-                gamePeriod = 1;
-            }
-            else
-            {
-                if (gamePeriod == null && checkPlay.GamePeriod != null)
-                {
-                    gamePeriod = checkPlay.GamePeriod;
-                }
-                if (checkPlay.ClockEnd != null)
-                {
-                    clockStart = checkPlay.ClockEnd;
-                }
-                else if (checkPlay.ClockStart != null)
-                {
-                    clockStart = checkPlay.ClockStart;
-                }
-                else if (gamePeriod != null && (checkPlay.GamePeriod ?? gamePeriod) != gamePeriod)
-                {
-                    clockStart = game.PeriodLength;
-                }
-                else
-                {
-                    playCheckId = checkPlay.PrevPlayId ?? 0;
-                }
-            }
-        } while (clockStart == null || gamePeriod == null);
-
-        if (clockStart == 0)
-        {
-            gamePeriod += 1;
-            clockStart = game.PeriodLength;
-        }
-
-        PlaySubmitDTO nextPlay = new()
-        {
-            PrevPlayId = currentPlay?.Id ?? -1,
-            GameId = game.Id,
-            TeamId = nextTeamId ?? 0,
-            FieldPositionStart = fieldPositionStart,
-            Down = down,
-            ToGain = toGain,
-            ClockStart = clockStart,
-            GamePeriod = gamePeriod
-        };
-
-        GameStreamDTO gameStream = new(game, nextPlay);
-
-        gameStream.HomeTeamScore = homeTeamScore;
-        gameStream.AwayTeamScore = awayTeamScore;
-        gameStream.LastPlay = lastPlay;
-
-        // The drive always has at least one play in it (that play may be an empty play if at start of game or currentPlay is otherwise null)
-        List<Play> drive = [currentPlay ?? new()];
-
-        bool driveFound = currentPlay?.TeamId != nextTeamId;
-
-        // Collect all plays from current drive, including kickoff to start drive (does not count as a play)
-        while (drive[0].Kickoff == null && drive[0].PrevPlayId > 0 && !driveFound)
-        {
-            Play? previousPlay = game.Plays.SingleOrDefault((p) => p.Id == drive[0].PrevPlayId);
-            if (previousPlay == null)
-            {
-                driveFound = true;
-            }
-            else if (previousPlay.TeamId != nextTeamId && previousPlay.Kickoff == null)
-            {
-                driveFound = true;
-            }
-            else
-            {
-                drive.Insert(0, previousPlay);
-            }
-        }
-
-        // Remove kickoffs, turnovers, empty plays
-        List<Play> countedPlays = drive
-            .Where((p) => p.Kickoff == null
-                && p.TeamId == nextPlay.TeamId
-                && p.PrevPlayId != null)
-            .ToList();
-
-        gameStream.DrivePlays = countedPlays.Count();
-        gameStream.DrivePositionStart = nextPlay.FieldPositionStart;
-        gameStream.DriveYards = 0;
-        gameStream.DriveTime = 0;
-
-        if (gameStream.DrivePlays > 0)
-        {
-            int driveTimeStart = 0;
-            int driveTimeEnd = (game.GamePeriods - (nextPlay.GamePeriod ?? 0)) * game.PeriodLength + (nextPlay.ClockStart ?? 0);
-
-            gameStream.DrivePositionStart = countedPlays[0].FieldPositionStart;
-
-            driveTimeStart = (game.GamePeriods - (countedPlays[0].GamePeriod ?? 0)) * game.PeriodLength + (countedPlays[0].ClockStart ?? 0);
-
-            gameStream.DriveTime = driveTimeStart - driveTimeEnd;
-            // If last play was a made field goal, or a turnover, count drive yards up to start of last play
-            if ((countedPlays[countedPlays.Count - 1].FieldGoal?.Good ?? false)
-                || (nextTeamId != countedPlays[countedPlays.Count -1].TeamId))
-            {
-                gameStream.DriveYards = ((gameStream.DrivePositionStart - countedPlays[countedPlays.Count - 1].FieldPositionStart) * (currentPlay?.TeamId == game?.HomeTeamId ? -1 : 1)) ?? 0;
-            }
-            else if (nextPlay.Down != 0)
-            {
-                gameStream.DriveYards = ((gameStream.DrivePositionStart - nextPlay.FieldPositionStart) * (currentPlay?.TeamId == game?.HomeTeamId ? -1 : 1)) ?? 0;
-            }
-            // else count to end
-            else
-            {
-                gameStream.DriveYards = ((gameStream.DrivePositionStart - countedPlays[countedPlays.Count - 1].FieldPositionEnd) * (currentPlay?.TeamId == game?.HomeTeamId ? -1 : 1)) ?? 0;
-            }
-        }
-        gameStream.DrivePlays = countedPlays.Where((p) => !p.Penalties.Any((pp) => pp.Enforced && pp.NoPlay)).Count();
-
-        return gameStream;
     }
 
     public async Task<ResponseDTO<Game>> UpdateGameAsync(GameSubmitDTO gameSubmit)
